requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
python-dotenv==1.0.0
requests==2.32.4
openai==0.27.8
twilio==8.10.0 